requirements.txt

fastapi
uvicorn
jinja2
python-multipart
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
python-jose
weasyprint